move page components and styling to containers folder to fix hydration error
