+++ webapp/frontend/src/components/ThemeRegistry.jsx
@@ -5,11 +5,13 @@ import { CacheProvider } from '@emotion/react';
 import { ThemeProvider, CssBaseline } from '@mui/material';
 import createEmotionCache from '../lib/emotionCache';
 import { createTheme } from '@mui/material/styles';
+import { useState } from 'react';
 
 const theme = createTheme(); // You can customize this
-const emotionCache = createEmotionCache();
+
 
 export default function ThemeRegistry({ children }) {
+  const [emotionCache] = useState(()=>createEmotionCache());
   return (
     <CacheProvider value={emotionCache}>
       <ThemeProvider theme={theme}>
